refactor: move shared helper func to utils package
